Removing dockerized hashicorp vault
