+++ acceptance-tests/build.gradle
@@ -53,6 +53,13 @@ dependencies {
 
 test.enabled = false
 
+def vaultBinary () {
+  if (osdetector.os == "windows") {
+    return "vault.exe"
+  }
+  return "vault"
+}
+
 task downloadHashicorpVaultDist(type: Download) {
   src {
     // see gradle.properties for Hashicorp Vault URL
@@ -78,6 +85,7 @@ task acceptanceTest(dependsOn: [
   group = 'verification'
 
   systemProperty 'vaultInstallDir', "${buildDir}"
+  systemProperty 'vaultBinary', vaultBinary()
 
   useJUnitPlatform()
   // toggle to show standard out and standard error of the test JVM(s) on the console
