+++ acceptance-tests/src/test/java/tech/pegasys/signers/secp256k1/tests/multikey/HashicorpHelpers.java
@@ -17,8 +17,6 @@ import static java.util.Collections.singletonMap;
 import tech.pegasys.signers.hashicorp.dsl.HashicorpNode;
 import tech.pegasys.signers.secp256k1.HashicorpSigningParams;
 
-import com.github.dockerjava.api.DockerClient;
-
 public class HashicorpHelpers {
 
   public static final String secretPath = "acceptanceTestSecretPath";
@@ -26,9 +24,8 @@ public class HashicorpHelpers {
   private static final String PRIVATE_KEY_HEX_STRING =
       "8f2a55949038a9610f50fb23b5883af3b4ecb3c3bb792cbcefbd1542c692be63";
 
-  public static HashicorpSigningParams createLoadedHashicorpVault(
-      final DockerClient docker, boolean useTls) {
-    final HashicorpNode hashicorpNode = HashicorpNode.createAndStartHashicorp(docker, useTls);
+  public static HashicorpSigningParams createLoadedHashicorpVault(boolean useTls) {
+    final HashicorpNode hashicorpNode = HashicorpNode.createAndStartHashicorp(useTls);
     hashicorpNode.addSecretsToVault(singletonMap(secretName, PRIVATE_KEY_HEX_STRING), secretPath);
 
     return new HashicorpSigningParams(hashicorpNode, secretPath, secretName);
