+++ acceptance-tests/src/test/java/tech/pegasys/signers/secp256k1/tests/multikey/HashicorpTlsBasedTomlLoadingAcceptanceTest.java
@@ -15,7 +15,6 @@ package tech.pegasys.signers.secp256k1.tests.multikey;
 import static org.assertj.core.api.Assertions.assertThat;
 import static tech.pegasys.signers.secp256k1.MultiKeyTomlFileUtil.createHashicorpTomlFileAt;
 
-import tech.pegasys.signers.hashicorp.dsl.DockerClientFactory;
 import tech.pegasys.signers.secp256k1.HashicorpSigningParams;
 
 import java.nio.file.Path;
@@ -34,8 +33,7 @@ class HashicorpTlsBasedTomlLoadingAcceptanceTest extends MultiKeyAcceptanceTestB
 
   @BeforeAll
   static void setUpBase() {
-    hashicorpNode =
-        HashicorpHelpers.createLoadedHashicorpVault(new DockerClientFactory().create(), true);
+    hashicorpNode = HashicorpHelpers.createLoadedHashicorpVault(true);
   }
 
   @Test
