+++ gradle.properties
@@ -2,3 +2,4 @@ org.gradle.jvmargs=-Xmx1g
 version=1.0.2-SNAPSHOT
 hashicorpVaultUrl.osx=https://releases.hashicorp.com/vault/1.4.3/vault_1.4.3_darwin_amd64.zip
 hashicorpVaultUrl.linux=https://releases.hashicorp.com/vault/1.4.3/vault_1.4.3_linux_amd64.zip
+hashicorpVaultUrl.windows=https://releases.hashicorp.com/vault/1.4.3/vault_1.4.3_windows_amd64.zip
